+++ demo/ContactManager/AspNetCore/Program.cs
@@ -1,3 +1,5 @@
+using ViewModelShell;
+
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Services.AddControllers(options =>
@@ -15,7 +17,7 @@ builder.Services.AddControllers(options =>
 var app = builder.Build();
 
 app.UseDefaultFiles();
-app.UseStaticFiles();
+app.UseVmsShellStaticFiles();
 
 app.MapControllers();